search, pictures features

+++ user-management/routes/web.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\userController;
+use App\Http\Controllers\DepartmentController;
 
 /*
 |--------------------------------------------------------------------------
@@ -18,20 +19,21 @@ Route::post('/login' , [userController::class , 'login']);
 
 Route::get('/login' , [userController::class , function(){
     return view('login');
-}]);
+}])->name('login');
 
 Route::get('/logout' , [userController::class , 'logout']);
 
 
-Route::get('/', [userController::class , 'dashboard'] );
+Route::get('/', [userController::class , 'dashboard'] )->middleware('auth');
 
-Route::post('/create-user' , [userController::class , 'createUser']);
-Route::get('/create-user' , [userController::class , function(){
-    return view('createUser');
-}]);
+Route::post('/create-user' , [userController::class , 'createUser'])->middleware('auth');
+Route::get('/create-user' , [userController::class , 'createUserView'])->middleware('auth');
 
 
-Route::get('/user/{id}' ,  [userController::class , 'getUser'] );
-Route::delete('/user/{id}' ,  [userController::class , 'deleteUser'] );
-Route::put('edit-user/{id}' ,  [userController::class , 'editUser'] );
-Route::get('edit-user/{id}' ,  [userController::class , 'editUserTemplate']);
+Route::get('/user/{id}' ,  [userController::class , 'getUser'] )->middleware('auth');
+Route::delete('/user/{id}' ,  [userController::class , 'deleteUser'] )->middleware('auth');
+Route::put('edit-user/{id}' ,  [userController::class , 'editUser'] )->middleware('auth');
+Route::get('edit-user/{id}' ,  [userController::class , 'editUserView'])->middleware('auth');
+
+Route::post('/create-department' , [DepartmentController::class , 'createDepartment'])->middleware('auth');
+Route::get('/create-department' , [DepartmentController::class , 'createDepartmentView'])->middleware('auth');
